# api/openapi.yaml
openapi: 3.1.0
info:
  title: Learning Center Reference API
  version: 0.4.0
  description: >-
    Contract for an independent portfolio reference implementation using fictional
    soccer-federation data. It is not an official federation service.
servers:
  - url: http://localhost:8080
    description: Local Docker demonstration
tags:
  - name: Operations
  - name: Learning
  - name: Compliance
paths:
  /health:
    get:
      operationId: getHealth
      tags: [Operations]
      summary: Check API and database readiness
      responses:
        "200":
          description: Ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
        "503":
          description: Database unavailable
  /v1/members/{memberId}/eligibility:
    get:
      operationId: getMemberEligibility
      tags: [Compliance]
      summary: Compute current participation eligibility from source facts
      parameters:
        - $ref: "#/components/parameters/MemberId"
      responses:
        "200":
          description: Derived eligibility decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Eligibility"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/members/{subject}/credentials:
    get:
      operationId: getMemberCredentials
      tags: [Compliance]
      summary: Serve a member's credential facts and derived eligibility to a trusted service
      description: >-
        Provider side of the learning-center.credentials.v1 contract. The caller is a
        service, not a person: it presents an OAuth2 client-credentials token that must
        carry the credentials:read scope, and no member row is resolved for it. The path
        subject is the member's identity-provider subject, URL-encoded. Eligibility and
        every valid flag are evaluated by this API at as_of; consumers must not re-derive
        them from the dates. Hold reasons, display names, and dates of birth are never
        included.
      security:
        - serviceBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Subject"
      responses:
        "200":
          description: Credential facts and derived eligibility for the subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberCredentials"
        "400":
          description: Empty, malformed, or over-long subject
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "401":
          description: Missing or invalid service token
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "403":
          description: Valid token without the credentials:read scope
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "404":
          description: No member has that subject
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
  /v1/courses:
    get:
      operationId: listCourses
      tags: [Learning]
      summary: List published courses for the authenticated learner
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Published course catalog
          content:
            application/json:
              schema:
                type: object
                required: [courses]
                properties:
                  courses:
                    type: array
                    items:
                      $ref: "#/components/schemas/Course"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/courses/{courseId}/enrollments:
    post:
      operationId: enrollInCourse
      tags: [Learning]
      summary: Idempotently enroll the authenticated learner
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CourseId"
      responses:
        "200":
          description: Existing enrollment returned after a retry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrollmentMutation"
        "201":
          description: Enrollment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrollmentMutation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/enrollments/{enrollmentId}/lessons/{lessonId}/complete:
    post:
      operationId: completeLesson
      tags: [Learning]
      summary: Idempotently append a lesson-completed event and refresh progress
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/EnrollmentId"
        - $ref: "#/components/parameters/LessonId"
      responses:
        "200":
          description: Current projected enrollment progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompletionMutation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: A sequential course has an unfinished earlier lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/me/dashboard:
    get:
      operationId: getLearnerDashboard
      tags: [Learning]
      summary: Load the authenticated learner's projected progress
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Learner and enrollment progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/admin/compliance:
    get:
      operationId: listCompliance
      tags: [Compliance]
      summary: List current derived eligibility for an authenticated administrator
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Synthetic member compliance roster
          content:
            application/json:
              schema:
                type: object
                required: [members]
                properties:
                  members:
                    type: array
                    items:
                      $ref: "#/components/schemas/ComplianceMember"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        OIDC JWT in production mode; explicit synthetic token in local demo mode.
    serviceBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        OAuth2 client-credentials access token from the shared provider, verified for
        signature, issuer, audience, and expiry, then authorised on the credentials:read
        scope (space-separated "scope" claim or "scp" array). In local demo mode the
        synthetic DEMO_SERVICE_TOKEN maps to that scope.
  parameters:
    MemberId:
      name: memberId
      in: path
      required: true
      schema: {type: string, format: uuid}
    Subject:
      name: subject
      in: path
      required: true
      description: URL-encoded identity-provider subject of the member, for example demo%7Clearner
      schema: {type: string, minLength: 1, maxLength: 255}
    CourseId:
      name: courseId
      in: path
      required: true
      schema: {type: string, format: uuid}
    EnrollmentId:
      name: enrollmentId
      in: path
      required: true
      schema: {type: string, format: uuid}
    LessonId:
      name: lessonId
      in: path
      required: true
      schema: {type: string, format: uuid}
  responses:
    BadRequest:
      description: Invalid identifier
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    Forbidden:
      description: Identity is not provisioned or lacks the required database role
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    NotFound:
      description: Resource does not exist
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: {type: string}
    Health:
      type: object
      required: [status, database]
      properties:
        status: {type: string, enum: [ok, degraded]}
        database: {type: string, enum: [ok, unreachable, not configured]}
    Eligibility:
      type: object
      required: [member_id, status, reason]
      properties:
        member_id: {type: string, format: uuid}
        status: {type: string, enum: [eligible, suspended, ineligible_lapsed]}
        reason: {type: string}
    Member:
      type: object
      required: [id, display_name, roles]
      properties:
        id: {type: string, format: uuid}
        display_name: {type: string}
        roles:
          type: array
          items: {type: string, enum: [learner, instructor, admin, coach, referee]}
    Course:
      type: object
      required: [id, title, slug, ordering, lesson_count]
      properties:
        id: {type: string, format: uuid}
        title: {type: string}
        slug: {type: string}
        ordering: {type: string, enum: [sequential, open]}
        lesson_count: {type: integer, minimum: 0}
    LessonProgress:
      type: object
      required: [id, title, type, position, completed]
      properties:
        id: {type: string, format: uuid}
        title: {type: string}
        type: {type: string, enum: [video, reading, quiz]}
        position: {type: integer, minimum: 1}
        completed: {type: boolean}
        completed_at: {type: string, format: date-time}
    EnrollmentProgress:
      type: object
      required:
        - enrollment_id
        - course_id
        - course_title
        - status
        - completed_lessons
        - total_lessons
        - percent_complete
        - lessons
      properties:
        enrollment_id: {type: string, format: uuid}
        course_id: {type: string, format: uuid}
        course_title: {type: string}
        status: {type: string, enum: [active, completed, withdrawn]}
        completed_lessons: {type: integer, minimum: 0}
        total_lessons: {type: integer, minimum: 0}
        percent_complete: {type: integer, minimum: 0, maximum: 100}
        lessons:
          type: array
          items: {$ref: "#/components/schemas/LessonProgress"}
    Dashboard:
      type: object
      required: [member, enrollments]
      properties:
        member: {$ref: "#/components/schemas/Member"}
        enrollments:
          type: array
          items: {$ref: "#/components/schemas/EnrollmentProgress"}
    EnrollmentMutation:
      type: object
      required: [created, enrollment]
      properties:
        created: {type: boolean}
        enrollment: {$ref: "#/components/schemas/EnrollmentProgress"}
    CompletionMutation:
      type: object
      required: [recorded, enrollment]
      properties:
        recorded: {type: boolean}
        enrollment: {$ref: "#/components/schemas/EnrollmentProgress"}
    ComplianceMember:
      type: object
      required: [id, display_name, roles, status, reason]
      properties:
        id: {type: string, format: uuid}
        display_name: {type: string}
        roles:
          type: array
          items: {type: string}
        status: {type: string, enum: [eligible, suspended, ineligible_lapsed]}
        reason: {type: string}
        next_expiration: {type: string, format: date-time}
    MemberCredentials:
      type: object
      description: >-
        learning-center.credentials.v1 response. Adding a field is compatible; removing or
        renaming one, or changing an enum, is a new contract version.
      required: [contract, member, as_of, eligibility, holds, safeguarding, role_credentials]
      properties:
        contract: {type: string, enum: [learning-center.credentials.v1]}
        member: {$ref: "#/components/schemas/CredentialsMember"}
        as_of: {type: string, format: date-time}
        eligibility: {$ref: "#/components/schemas/CredentialsEligibility"}
        holds:
          type: array
          items: {$ref: "#/components/schemas/ActiveHold"}
        safeguarding:
          type: object
          required: [safesport_training, background_check]
          properties:
            safesport_training: {$ref: "#/components/schemas/ExpiringFact"}
            background_check: {$ref: "#/components/schemas/ExpiringFact"}
        role_credentials:
          type: array
          items: {$ref: "#/components/schemas/RoleCredentialStatus"}
    CredentialsMember:
      type: object
      required: [id, subject, roles]
      properties:
        id: {type: string, format: uuid}
        subject: {type: string}
        roles:
          type: array
          items: {type: string, enum: [learner, instructor, admin, coach, referee]}
    CredentialsEligibility:
      type: object
      required: [status, reason]
      properties:
        status: {type: string, enum: [eligible, suspended, ineligible_lapsed]}
        reason: {type: string, description: Free text for people; never parsed by consumers}
    ActiveHold:
      type: object
      description: An active disciplinary hold; the reason text is deliberately not carried
      required: [source, active]
      properties:
        source: {type: string, enum: [us_soccer, safesport, local_org]}
        active: {type: boolean}
    ExpiringFact:
      type: object
      required: [expires_at, valid]
      properties:
        expires_at:
          type: [string, "null"]
          format: date
          description: Null when nothing is on file
        valid: {type: boolean, description: Evaluated by this API at as_of with its inclusive-expiry rule}
    RoleCredentialStatus:
      type: object
      required: [role, credential_type, issued_at, expires_at, valid]
      properties:
        role: {type: string, enum: [coach, referee]}
        credential_type: {type: string}
        issued_at: {type: string, format: date}
        expires_at: {type: string, format: date}
        valid: {type: boolean}